requirements.txt
-e .
aws_cdk.core
aws_cdk.aws_apigateway
aws_cdk.aws_dynamodb
aws_cdk.aws_ec2
aws_cdk.aws_iam
aws_cdk.aws_lambda
aws_cdk.aws_lambda_event_sources
aws_cdk.aws_s3
aws_cdk.aws_sns
aws_cdk.aws_sns_subscriptions
aws_cdk.aws_sqs
pytest